security detail fix
This bug manifested when the cardinality of the customers was increased.
The PostgreSQL stored function was returning the s_num_out as a 32-bit
integer, which caused errors when the scale of the customers increased
the possible values beyond that range.  The stored function definition
was fixed and the transaction code also has to be fixed to handle values
that would overrun a C int.

+++ code/src/transactions/pgsql/DBConnection.cpp
@@ -790,7 +790,7 @@ void CDBConnection::execute(const TSecurityDetailFrame1Input *pIn,
 			&pOut->open_date.day);
 	pOut->pe_ratio = c[i_pe_ratio].as(double());
 	strncpy(pOut->s_name, c[i_s_name].c_str(), cS_NAME_len);
-	pOut->num_out = c[i_num_out].as(int());
+	pOut->num_out = c[i_num_out].as(long());
 	strncpy(pOut->sp_rate, c[i_sp_rate].c_str(), cSP_RATE_len);
 	sscanf(c[i_start_date].c_str(), "%hd-%hd-%hd",
 			&pOut->start_date.year,
